requirements.txt
streamlit>=1.39.0
openai>=1.52.0
groq>=0.11.0
google-generativeai>=0.8.3
sentence-transformers>=3.2.0
faiss-cpu>=1.8.0
pdfplumber>=0.11.4
pandas>=2.2.3
numpy>=2.0.0
python-dotenv>=1.0.1
tavily-python>=0.5.0
matplotlib>=3.9.0
streamlit-extras>=0.5.0
plotly>=5.24.0
